techdocs-node: inline swift sdk mock
